Added support for ordered queries, while still maintaining dictionary style query parameters.
URLComponents' `queryItems` is an array meaning it is ordered, and therefore the resulting queries in the URL string are also ordered. While the order is not important for actual requests, it can be useful as properties such as URLs `absoluteString`, where it will now always maintain the same order of queries.

Frameworks for mocking URLSession requests, e.g. Mocker, catch URLs by using `absoluteString`, when deciding when to return mock data.